+++ exeris-sdk-annotations/src/main/java/eu/exeris/sdk/annotation/DomainEvent.java
@@ -837,6 +837,8 @@ public @interface DomainEvent {
     /**
      * Message header definition.
      */
+    @Retention(RetentionPolicy.SOURCE)
+    @Target({})
     @interface Header {
         /**
          * Header name.